Projet Ingénieur ECAM EPMI/CyberSécurité Projet/CyberSécurité_Devoir_Projets/Vs Code/Cybersécurité/Crypto/Application_Cours/requirements.txt
asttokens==2.4.1
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fonttools==4.47.0
idna==3.6
ipykernel==6.28.0
ipython==8.19.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.0
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.3
packaging==23.2
pandas==2.1.4
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pycryptodome==3.19.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-nmap==0.7.1
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
requests==2.31.0
scapy==2.5.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.4
traitlets==5.14.1
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
winpcapy==1.0.2
